Separate quality manager related classes into separate files
